Rename typo'd directories
