+++ Warcraft.NET/Files/ADT/Entries/Legion/MLMDEntry.cs
