+++ Warcraft.NET/Files/ADT/Entries/MoP/MTXPEntry.cs
