+++ Warcraft.NET/Files/ADT/TerrainTexture/MapChunk/Entries/MCLYEntry.cs
