+++ Warcraft.NET/Files/M2/Entries/AFIDEntry.cs
