+++ Warcraft.NET/Files/M2/Entries/MD21Entry.cs
