+++ Warcraft.NET/Files/TEX/Entries/TXBTEntry.cs
