+++ Warcraft.NET/Files/WDT/Entries/Entrys/BfA/MAIDEntry.cs
