+++ Warcraft.NET/Files/WDT/Entries/Entrys/Legion/MPL2Entry.cs
