+++ Warcraft.NET/Files/WDT/Entries/Entrys/MAINEntry.cs
