+++ Warcraft.NET/Files/WDT/Entries/Entrys/WoD/MAOIEntry.cs
